refactor(core): extract pure sanitizeForPath into naming module
